+++ baybe/parameters/__init__.py
@@ -2,6 +2,7 @@
 
 from baybe.parameters.categorical import CategoricalParameter, TaskParameter
 from baybe.parameters.custom import CustomDiscreteParameter
+from baybe.parameters.encoding import EncoderProtocol
 from baybe.parameters.enum import (
     CategoricalEncoding,
     CustomEncoding,
@@ -11,6 +12,7 @@ from baybe.parameters.numerical import (
     NumericalContinuousParameter,
     NumericalDiscreteParameter,
 )
+from baybe.parameters.sequence import SequenceParameter
 from baybe.parameters.substance import SubstanceParameter
 from baybe.utils.metadata import MeasurableMetadata
 
@@ -19,9 +21,11 @@ __all__ = [
     "CategoricalParameter",
     "CustomDiscreteParameter",
     "CustomEncoding",
+    "EncoderProtocol",
     "MeasurableMetadata",
     "NumericalContinuousParameter",
     "NumericalDiscreteParameter",
+    "SequenceParameter",
     "SubstanceEncoding",
     "SubstanceParameter",
     "TaskParameter",